requirements.txt
django
djangorestframework
django-cors-headers
pandas
PyQt5
matplotlib
requests
